# openapi.yaml
openapi: 3.0.3
info:
  title: Push Balancer API
  description: |
    ML-gestützte API für BILD Push-Notification-Optimierung.

    Der Push Balancer analysiert historische Push-Daten, prognostiziert Opening Rates (OR)
    via Gradient Boosted Trees und Stacking Ensemble (LightGBM + XGBoost + CatBoost → Ridge),
    und liefert tagesaktuelle Empfehlungen für Redaktion und Sport.

    **Safety**: Alle Vorhersagen sind ausschließlich beratend (ADVISORY_ONLY). Das System
    darf niemals autonom Push-Benachrichtigungen versenden. Jede Prediction enthält
    `advisoryOnly: true` und `actionAllowed: false`.
  version: 2.4.0
  contact:
    name: BILD Digital Produkt
    email: push-balancer@axelspringer.com
servers:
  - url: http://localhost:8050
    description: Lokaler Entwicklungsserver
  - url: https://push-balancer.onrender.com
    description: Render Cloud Deployment
...
paths:
...
  /api/feed:
    get:
      operationId: getBildSitemapFeed
      tags: [feeds]
      summary: BILD News-Sitemap abrufen
      description: |
        Proxy zu `bild.de/sitemap-news.xml`. Liefert die aktuelle BILD News-Sitemap
        als XML. Wird im Frontend zur Artikel-Auswahl für Push-Kandidaten genutzt.
      responses:
        "200":
          description: BILD News-Sitemap als XML
          content:
            application/xml:
              schema:
                type: string
                format: binary
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/competitors:
    get:
      operationId: getCompetitorFeeds
      tags: [feeds]
      summary: Alle Competitor-Feeds abrufen
      description: |
        Liefert geparsed RSS-Feeds aller deutschen Nachrichtenkonkurrenten aus dem
        Background-Cache (Welt, Spiegel, Focus, n-tv, Tagesschau, FAZ, SZ, Stern,
        t-online, Zeit). Response-Zeit unter 5 ms da gecacht.
      responses:
        "200":
          description: Alle Competitor-Feeds als geparste JSON-Objekte
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedCollection"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/sport-competitors:
    get:
      operationId: getSportCompetitorFeeds
      tags: [feeds]
      summary: Sport-Competitor-Feeds abrufen
      description: |
        Liefert geparsed RSS-Feeds der Sport-Konkurrenten (Kicker, Sportschau,
        Transfermarkt etc.) aus dem Background-Cache.
      responses:
        "200":
          description: Sport-Competitor-Feeds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedCollection"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/sport-europa:
    get:
      operationId: getSportEuropaFeeds
      tags: [feeds]
      summary: Europäische Sport-Feeds abrufen
      description: |
        Liefert geparsed RSS-Feeds europäischer Sport-Outlets aus dem Background-Cache.
      responses:
        "200":
          description: Europäische Sport-Feeds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedCollection"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/sport-global:
    get:
      operationId: getSportGlobalFeeds
      tags: [feeds]
      summary: Globale Sport-Feeds abrufen
      description: |
        Liefert geparsed RSS-Feeds globaler Sport-Outlets (internationale Ligen,
        Weltklasse-Medien) aus dem Background-Cache.
      responses:
        "200":
          description: Globale Sport-Feeds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedCollection"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/international:
    get:
      operationId: getInternationalFeeds
      tags: [feeds]
      summary: Internationale News-Feeds abrufen
      description: |
        Liefert geparsed RSS-Feeds von 24 internationalen Nachrichtenquellen aus
        Europa (BBC, Guardian, Le Monde, El País usw.) und Global (CNN, NYT,
        Al Jazeera usw.) aus dem Background-Cache.
      responses:
        "200":
          description: Internationale News-Feeds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedCollection"
        "502":
          $ref: "#/components/responses/BadGateway"
...
  /api/competitor-xor:
    post:
      operationId: competitorXor
      tags: [analytics]
      summary: Competitor-Artikel OR-Prognose (Batch)
      description: |
        Berechnet schnelle Opening-Rate-Prognosen für Competitor-Artikel-Titel via
        Wort-Performance-Scoring (O(W)-Lookup statt O(N×M)-Jaccard). Bis zu 200
        Titel in einem Request, Antwort in unter 50 ms. Erkennt automatisch
        Kategorien, Eilmeldungen und Live-Ticker (niedrigere Konfidenz).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetitorXorRequest"
      responses:
        "200":
          description: OR-Prognosen für alle Titel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitorXorResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 Problem Details für HTTP APIs
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: URI-Referenz, die den Fehlertyp identifiziert
          example: "https://push-balancer.example.com/problems/validation-error"
        title:
          type: string
          description: Kurze, menschenlesbare Fehlerbeschreibung
          example: "Validation Error"
        status:
          type: integer
          description: HTTP-Statuscode
          example: 400
        detail:
          type: string
          description: Ausführlichere Fehlerbeschreibung
          example: "Das Feld 'pushId' darf nicht leer sein"
        instance:
          type: string
          format: uri
          description: URI-Referenz des konkreten Fehler-Vorkommens
          example: "/api/prediction-feedback"

    # ── Health ──────────────────────────────────────────────────────────────
...
    RssItem:
      type: object
      description: Einzelnes RSS-Feed-Item (geparst)
      properties:
        title:
          type: string
          description: Artikeltitel
          example: "Merz zum Kanzler gewählt"
        link:
          type: string
          format: uri
          description: Artikel-URL
          example: "https://www.welt.de/artikel/1234"
        pubDate:
          type: string
          description: Veröffentlichungsdatum (RFC 2822)
          example: "Thu, 02 Apr 2026 14:30:00 +0200"
        description:
          type: string
          description: Teaser-Text
    FeedCollection:
      type: object
      description: Map von Outlet-Name zu Liste von RSS-Items
      additionalProperties:
        type: array
        items:
          $ref: "#/components/schemas/RssItem"
      example:
        welt:
          - title: "Eilmeldung: Erdbeben"
            link: "https://www.welt.de/..."
            pubDate: "Thu, 02 Apr 2026 10:00:00 +0200"

    # ── Research Rules ──────────────────────────────────────────────────────
...
    CompetitorXorRequest:
      type: object
      required: [titles]
      properties:
        titles:
          type: array
          description: Titel-Liste (max. 200; kann Strings oder Objekte sein)
          maxItems: 200
          items:
            oneOf:
              - type: string
                description: Einfacher Titelstring
              - type: object
                description: Titel mit Kategorie und Stunde
                properties:
                  title:
                    type: string
                  cat:
                    type: string
                  hour:
                    type: integer
    CompetitorXorEntry:
      type: object
      properties:
        or:
          type: number
          format: float
          description: Prognostizierte OR
          example: 5.84
        q10:
          type: number
          format: float
          example: 3.2
        q90:
          type: number
          format: float
          example: 8.9
        confidence:
          type: number
          format: float
          example: 0.54
        basis:
          type: string
          example: "word(6)"
        cat:
          type: string
          example: Politik
        nWords:
          type: integer
          description: Anzahl gescorefer Wörter
          example: 6
        isLiveticker:
          type: boolean
          description: Wurde der Titel als Live-Ticker erkannt?
          example: false
    CompetitorXorResponse:
      type: object
      properties:
        predictions:
          type: object
          description: Map von Titel (max. 120 Zeichen) zu Prognose
          additionalProperties:
            $ref: "#/components/schemas/CompetitorXorEntry"
        count:
          type: integer
          example: 42
        elapsedMs:
          type: number
          example: 38.7

  # ── Responses ────────────────────────────────────────────────────────────
  responses:
    BadRequest:
      description: Ungültige Anfrage (fehlende oder falsche Parameter)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "https://push-balancer.example.com/problems/bad-request"
            title: "Bad Request"
            status: 400
            detail: "Das Feld 'pushId' darf nicht leer sein"
...
    InternalServerError:
      description: Interner Serverfehler
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "https://push-balancer.example.com/problems/internal-error"
            title: "Internal Server Error"
            status: 500
            detail: "ML status error: model not initialized"
    BadGateway:
      description: Upstream-Fehler (externe API nicht erreichbar)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "https://push-balancer.example.com/problems/bad-gateway"
            title: "Bad Gateway"
            status: 502
            detail: "Proxy error: Connection timeout"